requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
cloudscraper==1.2.64
idna==3.4
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-toolbelt==0.10.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
